requirements.txt
django==5.2.3
djangorestframework==3.15.1
